+++ MantleArrayExtension/Classes/MAEFragment.m
@@ -78,4 +78,45 @@ extern MAEFragment* _Nonnull MAEVariadic(id _Nonnull v)
     return self;
 }
 
+#pragma mark - NSObject (Override)
+
+- (BOOL)isEqual:(id _Nullable)other
+{
+    if ([other isKindOfClass:MAEFragment.class]) {
+        typeof(self) otherFragment = other;
+        return [otherFragment.propertyName isEqual:self.propertyName]
+            && otherFragment.type == self.type
+            && otherFragment.optional == self.optional
+            && otherFragment.variadic == self.variadic;
+    }
+    return NO;
+}
+
+- (NSUInteger)hash
+{
+    return [self.propertyName hash];
+}
+
+- (NSString* _Nonnull)description
+{
+    char type = ' ';
+    switch (self.type) {
+        case MAEFragmentEnumerateString:
+            type = 'E';
+            break;
+        case MAEFragmentDoubleQuotedString:
+            type = 'D';
+            break;
+        case MAEFragmentSingleQuotedString:
+            type = 'S';
+            break;
+        default:
+            type = '-';
+    }
+
+    return [NSString stringWithFormat:@"<%@: %@ :%c%c%c>",
+                                      self.class, self.propertyName, type,
+                                      (self.optional ? 'O' : '-'), (self.variadic ? 'V' : '-')];
+}
+
 @end