+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Support Carthage.
 - Support some primitive types and NSNumber with default transformer.
+- The validation of formatByPropertyKey has been enhanced.
 
 ### Minor changes